\name{Internal functions}
\alias{arrayScal}
\alias{quality2HTML}
\alias{EMSplit}
\alias{qpS2N}
\alias{qpPTLoess}
\alias{qpImage}
\alias{qpHexbin}
\alias{gpFlagWt}
\alias{qpDotPlots}
\alias{qpMAPlots}
\alias{setCtlCol}
\alias{arrayControls}
\alias{replicatesAvariance}
\alias{arrayReplicates}
\alias{outputNormData}

\alias{qpS2Nmeebo}
\alias{qpDotPlotsMeebo}
\alias{qpDotPlotsMeebo2}
\alias{qpBoxplotMeebo}
\alias{controlCodeMeebo}
\alias{qpMisMatchPlot}
\alias{BEbin.linear}
\alias{qpBEplot.linear}
\alias{qpTiling}
\alias{getSpikeIds}
\alias{getSpikeIndex}
\alias{readAllSpikes}
\alias{Spike.MMplot}
\alias{Spike.Sensitivity}
\alias{Spike.Individual.Sensitivity}
\alias{Spike.MM.Scatter}
\alias{Spike.Cy5vsCy3}
\alias{MeeboSpotTypes}

\alias{qpDotPlotsEEBO}
\alias{qpS2Neebo}
\alias{controlCodeHeebo}
\alias{HeeboSpotTypes}


\title{Internal arrayQuality functions}
 
\description{Internal arrayQuality functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
