\name{gprDB}
\alias{MmScalingTable}
\alias{MmReferenceDB}
\alias{index.html}

\title{Reference slides for Mouse oligos hybridizations}

\description{
MmReferenceDB: quality matrix compiling results from a pool of
good hybridization slides.

MmScalingTable: matrix of means and iqr calculated from the pool of good
hybridization slides, for each quality measure. It is used to scale
other arrays quality results.

index.html: HTML file used for quality report.
}

\source{These data were provided by members of the UCSF Shared
  functional genomics core lab.}

\keyword{datasets}

