\name{MmDEGenes}
\alias{MmDEGenes}

\title{Known DE genes for Mouse quality hybridizations.}

\description{
  MmDEGenes contains information about probes known to be DE from
  previous quality hybridizations. It is used to verify reproducibility
  of print-runs.
}

\source{These data were provided by members of the UCSF Shared
  functional genomics core lab.}

\keyword{datasets}
