\name{prdata}
\alias{9Mm137.gpr}
\alias{12Mm250.gpr}
\alias{RDI108\_n.gpr}
\alias{StanfordDCV1.7complete.txt}
\alias{MeeboSpotType.txt}
\title{Example GPR files}

\description{
  9Mm137.gpr is QCHyb from 9Mm printrun.
  
  12Mm250.gpr is a 9mers hybridization from 12Mm printrun.
  
  RDI108\_n.gpr is an example of hybridization on MEEBO arrays, using a
  doping-control mixture from SFGF.
  StanfordDCV1.7complete.txt is an example of spike-type file from SFGF.
  
  MeeboSpotType.txt is an example of Spot Types File for the MEEBO set.
}

\source{These data were provided by members of the UCSF Shared
  Functional Genomics Core Facility, the Brown lab (Stanford) and the
  Stanford Shared Genomics Core Facility.}

\keyword{datasets}




