\name{Internal functions}
\alias{arrayScal}
\alias{qualBoxplot}
\alias{qualRefTable}
\alias{quality2HTML}
\alias{maQualityPlots}
\alias{maCompLayout}
\alias{EMSplit}
\alias{qpS2N}
\alias{qpPTLoess}
\alias{qpImage}
\alias{qpHexbin}
\alias{setCtlCol}
\alias{arrayControls}
\alias{replicatesAvariance}

\title{Internal arrayQuality functions}
 
\description{Internal arrayQuality functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
