\name{Internal functions}
\alias{arrayScal}
\alias{EMSplit}
\alias{globalQuality}
\alias{qualPlot}
\alias{qualRefTable}
\alias{readGPR}
\alias{slideQuality}
\alias{qpS2N}
\alias{qpPTLoess}
\alias{qpImage}
\alias{qpHexbin}
\alias{setCtlCol}

\title{Internal arrayQuality functions}
 
\description{Internal arrayQuality functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
