\name{Internal functions}
\alias{arrayScal}
\alias{quality2HTML}
\alias{EMSplit}
\alias{qpS2N}
\alias{qpPTLoess}
\alias{qpImage}
\alias{qpHexbin}
\alias{gpFlagWt}
\alias{qpDotPlots}
\alias{qpMAPlots}
\alias{setCtlCol}
\alias{arrayControls}
\alias{replicatesAvariance}
\alias{arrayReplicates}
\alias{outputNormData}

\title{Internal arrayQuality functions}
 
\description{Internal arrayQuality functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
