\name{prdata}
\alias{9Mm137.gpr}
\alias{12Mm250.gpr}
\title{Example GPR files}

\description{
  9Mm137.gpr is QCHyb from 9Mm printrun.
  
  12Mm250.gpr is a 9mers hybridization from 12Mm printrun.
}

\source{These data were provided by members of the UCSF Shared
  Functional Genomics Core Facility.}

\keyword{datasets}




